+++ app/admin/datafile.rb
@@ -1,4 +1,4 @@
-ActiveAdmin.register Datafile do
+ActiveAdmin.register Link do
 # See permitted parameters documentation:
 # https://github.com/activeadmin/activeadmin/blob/master/docs/2-resource-customization.md#setting-up-strong-parameters
 #
