+++ db/migrate/20180116132324_rename_datafiles_to_links.rb
@@ -1,0 +1,5 @@
+class RenameDatafilesToLinks < ActiveRecord::Migration[5.1]
+  def change
+    rename_table :datafiles, :links
+  end
+end
