+++ lib/tasks/import.rake
@@ -76,7 +76,7 @@ namespace :import do
 
   desc "Import datasets from a data.gov.uk dump"
   task :legacy_datasets, [:filename] => :environment do |_, args|
-    Link.skip_callback(:save, :before, :set_end_date)
+    Datafile.skip_callback(:save, :before, :set_end_date)
 
     # Maps the organisation UUIDs to the organisation IDs
     logger = Logger.new(STDOUT)
